{"schema":"decodex/commit/1","summary":"Extract host FFI session bridge owner","authority":"manual"}
